% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ludwig_functions.R
\name{create_matrix}
\alias{create_matrix}
\title{Create the predictor matrix and the vector of 
observed node states for pseudolikelihood estimation}
\usage{
create_matrix(data)
}
\arguments{
\item{data}{a binary data matrix containing observed network states. 
The rows represent the states, the columns represent the
nodes.}
}
\value{
A list containing... 
  \itemize{
  \item {\code{data}} {the original data matrix}.
  \item {\code{predict_matrix}} {the predictor matrix}.
  \item {\code{y}} {a vector of the observed node states}.
  \item {\code{n}} {the number of observed state vectors}.
  \item {\code{k}} {the number of nodes}.
  \item {\code{n_links}} {the number of links in the network}.
  \item {\code{n_param}} {the number of parameters to be estimated}.
  }
}
\description{
The main purpose of this function is to create 
    a predictor matrix (\code{predict_matrix}) and a vector of observed states 
    (\code{y}) for each node
    so that the network parameters are estimable using simple
    multivariate logistic regression based on the conditional distributions
    of each node given the rest of the network. This method is also 
    known as the pseudolikelihood method, supposedly first introduced by Besag (1975).
    Note that the speed of this function is improvable.
}
