% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ludwig_functions.R
\name{crossval}
\alias{crossval}
\title{Cross-validate network on new data}
\usage{
crossval(net, dat_val)
}
\arguments{
\item{net}{An object of class estnet}

\item{dat_val}{A validation dataset}
}
\value{
The relative frequency of correct predictions.
}
\description{
This intended use of this function is helping cross-validate
    an estimated network on new data. To this aim, the values of individual
    nodes are predicted based on the conditional predictive distributions
    and the relative frequency of correct predictions is returned.
}
