% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ludwig_functions.R
\name{diag_plot}
\alias{diag_plot}
\title{Diagnostic plots}
\usage{
diag_plot(x, ...)
}
\arguments{
\item{x}{an object of class \code{estnet_bayes}}

\item{...}{other arguments passed over to \code{mcmcplot}}
}
\description{
This function is a wrapper around \code{mcmcplots} to be found 
    in the \code{coda} package. It allows for a convenient visual 
    assessment of the Markov chains and the autocorrelation in a Web browser.
}
\seealso{
\code{autocorr.diag}, \code{coda}
}
