% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ludwig_functions.R
\name{DIC}
\alias{DIC}
\title{Deviance information criterion}
\usage{
DIC(x, ...)
}
\arguments{
\item{x}{an object of class \code{estnet_bayes}}

\item{...}{other arguments passed over to \code{dic.samples()}}
}
\description{
This function is a wrapper around \code{rjags::dic.samples()} to be found 
    in the \code{rjags} package. It allows for assessing the 
    deviance information criterion.
}
\seealso{
\code{dic.samples}, \code{coda}
}
