% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ludwig_functions.R
\name{estnet}
\alias{estnet}
\title{Estimate the network parameters given an observed state matrix}
\usage{
estnet(data, method = "glmnet", alpha = 0.5, s = "lambda.min", ...)
}
\arguments{
\item{data}{a binary matrix of observed network states. The rows are the states, the columns are the nodes.}

\item{method}{the method to be used. Either \code{"glmnet"} (default) or \code{"glm"}.}

\item{alpha}{the elastic mixing parameter passed to \code{\link[glmnet]{cv.glmnet}}.}

\item{s}{value of the penalty parameter passed to \code{\link[glmnet]{predict.cv.glmnet}}.}

\item{...}{other arguments passed over to e.g. \code{\link[glmnet]{cv.glmnet}}.}
}
\value{
The function returns a list of class \code{\link{estnet}}, containing...
  \itemize{
  \item {the objects returned by the function \code{\link{create_matrix}}}.
  \item {\code{fitted}} { fitted values based on the estimated parameters.}
  \item {\code{y}} { a vector of the observed node states.}
  \item {\code{logL}} { the log pseudo likelihood.}
  \item {\code{deviance}} { the deviance.}
  \item {\code{coefficients}} { the estimated coefficients (thresholds and weights).}
  \item {\code{weights}} { the estimated weights.}
  \item {\code{thresholds}} { the estimated thresholds.}
  \item {\code{time}} { the time used for the estimation process.}
  \item {\code{net}} { the list returned by either \code{\link[stats]{glm}} or \code{\link[glmnet]{cv.glmnet}}.}
  }
}
\description{
This function accepts a binary matrix of network states and estimates
   the thresholds and weights of the network model.
   The internally used functions are either \code{\link[glmnet]{cv.glmnet}} or \code{\link[stats]{glm}}.
   By default, \code{\link[glmnet]{cv.glmnet}} is used. 
   In this case, the mixing parameter \code{alpha} is set to \code{0.5} 
   (elastic net regularization) and \code{lambda} 
   is chosen based on the solution with mimimum mean cross-validation error 
   (argument \code{s="lambda.min"} passed to function \code{\link[glmnet]{predict.cv.glmnet}}).
}
\seealso{
\code{\link[glmnet]{glmnet}}, \code{\link[glmnet]{cv.glmnet}}, \code{\link[glmnet]{predict.cv.glmnet}},
    \code{\link[stats]{glm}}
}
