% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ludwig_functions.R
\name{estnet_bayes}
\alias{estnet_bayes}
\title{Determine the posterior distributions of the network parameters given an observed state matrix}
\usage{
estnet_bayes(data, n_chains = 2, n_iter = 1000, n_burnin = 1000,
  n_adapt = 1000, n_thin = 4)
}
\arguments{
\item{data}{a binary input matrix}

\item{n_chains}{the number of chains used}

\item{n_iter}{the number of MCMC iterations}

\item{n_burnin}{the number of burn-in iterations}

\item{n_adapt}{the number of adaption iterations}

\item{n_thin}{the thinning interval}
}
\value{
The function returns a list of class \code{\link{estnet_bayes}}, containing...
  \itemize{
  \item {the objects returned by the function \code{\link{create_matrix}}}
  \item {\code{s_coda}} {the raw coda samples}
  \item {\code{fitted}} { expected values of the nodes given the posterior means of the network parameters}
  \item {\code{entropy_rate}} {the entropy rate based on the network parameters' posterior means}
  \item {\code{logL}} { the log pseudo likelihood}
  \item {\code{deviance}} { the deviance}
  \item {\code{merged chains}} { the merged MCMC chains}
  \item {\code{used_samples}} { the actual number of samples to assess the posterior distributions}
  \item {\code{posterior_means}} { the parameter's posterior means}
  \item {\code{posterior_sd}} { the parameter's posterior standard deviations}
  \item {\code{weights}} { a matrix of the network parameter's posterior means}
  \item {\code{weights_sd}} { a matrix of the network parameter's posterior standard deviations}
  \item {\code{weights}} { the estimated weights}
  \item {\code{thresholds}} { posterior means of the threshold parameters}
  \item {\code{thresholds_sd}} { posterior standard deviations of the threshold parameters}
  \item {\code{n_chains}} {number of MCMC chains}
  \item {\code{thinning}} {thinning interval}
  \item {\code{burnin}} {burnin iterations}
  \item {\code{n_iter}} {number of iterations}
  \item {\code{actual}} {actual iterations used to compute the posterior statistics}
  \item {\code{time}} { the time used for the estimation process}
  }
}
\description{
This function allows for estimating an undirected graphical model
using MCMC via JAGS (Just another Gibb's sampler).
}
\examples{
# Simulates two negatively coupled nodes
# using random updating
simnet(2, c(1,1,-2), 1000, 1000)
}
