% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ludwig_functions.R
\docType{data}
\name{log_reg}
\alias{log_reg}
\title{Basic logistic regression model}
\format{An object of class \code{character} of length 1.}
\usage{
log_reg
}
\description{
Basic logistic regression model
}
\keyword{datasets}
