% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ludwig_functions.R
\name{plot.estnet}
\alias{plot.estnet}
\title{Plot the estimated network}
\usage{
\method{plot}{estnet}(x, ...)
}
\arguments{
\item{x}{an object of class \code{estnet}}

\item{...}{other arguments passed over to \code{qgraph}}
}
\description{
This function plots the estimated network.
    Basically, it is just a simple wrapper around  \code{\link[qgraph]{qgraph}},
    so all of qgraph's beautiful functionality should be available.
}
\seealso{
\code{\link[qgraph]{qgraph}}
}
