% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ludwig_functions.R
\name{plot.estnet_bayes}
\alias{plot.estnet_bayes}
\title{Plot the estimated network}
\usage{
\method{plot}{estnet_bayes}(x, thresh = 0, ...)
}
\arguments{
\item{x}{an object of class \code{estnet_bayes}}

\item{thresh}{a threshold in terms of posterior mean divided by posterior sd. 
For instance, setting the threshold to 2 removes all edges for which the 
absolute qoutient of posterior mean and posterior sd is smaller than 2.
This is useful for removing edges with small parameter values.}

\item{...}{other arguments passed over to \code{qgraph}}
}
\description{
This function plots the estimated network.
    Basically, it is just a simple wrapper around  \code{\link[qgraph]{qgraph}},
    so all of qgraph's functionality should be available.
}
\seealso{
\code{\link[qgraph]{qgraph}}
}
