% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ludwig_functions.R
\name{print.crossval}
\alias{print.crossval}
\title{Print the results}
\usage{
\method{print}{crossval}(x, ...)
}
\arguments{
\item{x}{An object of class \code{crossval}}

\item{...}{other arguments (not used)}
}
\description{
This function prints the results of the cross-validation procedure.
}
