% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ludwig_functions.R
\name{print.estnet}
\alias{print.estnet}
\title{Print the results}
\usage{
\method{print}{estnet}(netobj)
}
\arguments{
\item{netobj}{An object of class \code{estnet}}
}
\description{
This function prints the weights and
    thresholds of the estimated network as well as 
    other useful information.
}
