% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ludwig_functions.R
\name{print.estnet_bayes}
\alias{print.estnet_bayes}
\title{Print the results}
\usage{
\method{print}{estnet_bayes}(x, ...)
}
\arguments{
\item{x}{An object of class \code{estnet_bayes}}

\item{...}{other arguments (not used)}
}
\description{
This function prints the weights and
    thresholds of the estimated network as well as 
    other useful information.
}
