% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ludwig_functions.R
\name{R_hat}
\alias{R_hat}
\title{Determine the convergence of the Markov chains}
\usage{
R_hat(x, ...)
}
\arguments{
\item{x}{an object of class \code{estnet_bayes}}

\item{...}{other arguments passed over to \code{qgraph}}
}
\description{
This function is a wrapper around \code{gelman.diag} to be found 
    in the \code{coda} package. It computes the potential
    scale reduction for the Markov chains. This statistic is needed for the 
    assessment of model convergence.
}
\seealso{
\code{gelman.diag}, \code{coda}
}
